moved ui element context files to ui directory
