+++ fission/src/main.tsx
@@ -1,5 +1,5 @@
 import ReactDOM from "react-dom/client"
-import { Theme, ThemeProvider } from "./ThemeContext"
+import { Theme, ThemeProvider } from "@/ui/ThemeContext"
 import Synthesis from "./Synthesis"
 import "./index.css"
