+++ fission/src/ui/components/MainHUD.tsx
@@ -6,11 +6,11 @@ import { GrFormClose } from "react-icons/gr"
 import { GiSteeringWheel } from "react-icons/gi"
 import { HiDownload } from "react-icons/hi"
 import { IoGameControllerOutline, IoPeople } from "react-icons/io5"
-import { useModalControlContext } from "@/ModalContext"
-import { usePanelControlContext } from "@/PanelContext"
+import { useModalControlContext } from "@/ui/ModalContext"
+import { usePanelControlContext } from "@/ui/PanelContext"
 import { motion } from "framer-motion"
 import logo from "@/assets/autodesk_logo.png"
-import { ToastType, useToastContext } from "@/ToastContext"
+import { ToastType, useToastContext } from "@/ui/ToastContext"
 import { Random } from "@/util/Random"
 import APS, { APS_USER_INFO_UPDATE_EVENT } from "@/aps/APS"
 import { UserIcon } from "./UserIcon"
