+++ fission/src/ui/modals/configuring/SettingsModal.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { useModalControlContext } from "@/ModalContext"
+import { useModalControlContext } from "@/ui/ModalContext"
 import Modal, { ModalPropsImpl } from "@/components/Modal"
 import { FaGear } from "react-icons/fa6"
 import Label, { LabelSize } from "@/components/Label"
