+++ fission/src/ui/modals/configuring/rio-config/RCConfigEncoderModal.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react"
 import Modal, { ModalPropsImpl } from "@/components/Modal"
-import { useModalControlContext } from "@/ModalContext"
+import { useModalControlContext } from "@/ui/ModalContext"
 import { FaPlus } from "react-icons/fa6"
 import Label, { LabelSize } from "@/components/Label"
 import Input from "@/components/Input"
